+++ backend/de.metas.business/src/main/java/org/adempiere/mm/attributes/api/impl/ModelProductDescriptorExtractorUsingAttributeSetInstanceFactory.java
@@ -1,7 +1,7 @@
 package org.adempiere.mm.attributes.api.impl;
 
 import org.adempiere.mm.attributes.AttributeSetInstanceId;
-import org.adempiere.mm.attributes.api.AttributesKeys;
+import org.adempiere.mm.attributes.keys.AttributesKeys;
 import org.adempiere.mm.attributes.api.IAttributeSetInstanceAware;
 import org.adempiere.mm.attributes.api.IAttributeSetInstanceAwareFactoryService;
 import org.springframework.stereotype.Service;
