+++ backend/de.metas.business/src/main/java/org/adempiere/mm/attributes/keys/AttributeKeyPartPatternType.java
@@ -1,36 +1,38 @@
-package de.metas.material.commons.attributes;
-
-import java.util.Optional;
-
-import de.metas.material.event.commons.AttributesKey;
-
 /*
  * #%L
- * metasfresh-material-commons
+ * de.metas.business
  * %%
- * Copyright (C) 2019 metas GmbH
+ * Copyright (C) 2023 metas GmbH
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
-public interface AttributesKeyMatcher
+package org.adempiere.mm.attributes.keys;
+
+public enum AttributeKeyPartPatternType
 {
-	boolean matches(AttributesKey attributesKey);
+	All,
 
-	Optional<AttributesKey> toAttributeKeys();
+	Other,
 
-	AttributesKey toAttributeKeys(final AttributesKey context);
+	None,
+
+	AttributeValueId,
+
+	AttributeIdAndValue,
+
+	AttributeId;
 }
