+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/productplanning/PP_Product_Planning_StepDef.java
@@ -43,7 +43,7 @@ import lombok.NonNull;
 import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.ad.dao.IQueryBuilder;
 import org.adempiere.mm.attributes.AttributeSetInstanceId;
-import org.adempiere.mm.attributes.api.AttributesKeys;
+import org.adempiere.mm.attributes.keys.AttributesKeys;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.compiere.model.I_AD_Workflow;
 import org.compiere.model.I_C_UOM;
