+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/material/interceptor/transactionevent/HUDescriptorService.java
@@ -35,7 +35,7 @@ import de.metas.material.event.commons.ProductDescriptor;
 import de.metas.util.Services;
 import lombok.NonNull;
 import org.adempiere.mm.attributes.AttributeSetInstanceId;
-import org.adempiere.mm.attributes.api.AttributesKeys;
+import org.adempiere.mm.attributes.keys.AttributesKeys;
 import org.adempiere.mm.attributes.api.IAttributeSet;
 import org.adempiere.mm.attributes.api.IAttributeSetInstanceBL;
 import org.adempiere.mm.attributes.api.ImmutableAttributeSet;
