+++ backend/de.metas.manufacturing/src/main/java/org/eevolution/productioncandidate/service/PPOrderCandidatePojoConverter.java
@@ -41,7 +41,7 @@ import de.metas.util.Services;
 import lombok.NonNull;
 import org.adempiere.ad.persistence.ModelDynAttributeAccessor;
 import org.adempiere.mm.attributes.AttributeSetInstanceId;
-import org.adempiere.mm.attributes.api.AttributesKeys;
+import org.adempiere.mm.attributes.keys.AttributesKeys;
 import org.adempiere.warehouse.WarehouseId;
 import org.compiere.model.I_C_UOM;
 import org.eevolution.api.BOMComponentType;
