+++ backend/de.metas.material/cockpit/src/main/java/de/metas/material/cockpit/stock/StockRepository.java
@@ -4,8 +4,8 @@ import com.google.common.collect.ImmutableList;
 import de.metas.material.cockpit.model.I_MD_Stock;
 import de.metas.material.cockpit.model.I_MD_Stock_WarehouseAndProduct_v;
 import de.metas.material.cockpit.model.I_T_MD_Stock_WarehouseAndProduct;
-import de.metas.material.commons.attributes.AttributesKeyPatternsUtil;
-import de.metas.material.commons.attributes.AttributesKeyQueryHelper;
+import org.adempiere.mm.attributes.keys.AttributesKeyPatternsUtil;
+import org.adempiere.mm.attributes.keys.AttributesKeyQueryHelper;
 import de.metas.material.event.commons.AttributesKey;
 import de.metas.product.ProductId;
 import de.metas.util.Check;
