+++ backend/de.metas.material/planning/src/main/java/de/metas/material/planning/ppordercandidate/PPOrderCandidatePojoSupplier.java
@@ -49,7 +49,7 @@ import de.metas.util.Services;
 import lombok.NonNull;
 import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.mm.attributes.AttributeSetInstanceId;
-import org.adempiere.mm.attributes.api.AttributesKeys;
+import org.adempiere.mm.attributes.keys.AttributesKeys;
 import org.adempiere.service.ClientId;
 import org.compiere.model.I_M_Product;
 import org.eevolution.api.IProductBOMDAO;
