+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/order/products_proposal/service/OrderProductProposalsService.java
@@ -23,7 +23,7 @@ import de.metas.product.ProductId;
 import de.metas.util.Services;
 import lombok.NonNull;
 import org.adempiere.mm.attributes.AttributeSetInstanceId;
-import org.adempiere.mm.attributes.api.AttributesKeys;
+import org.adempiere.mm.attributes.keys.AttributesKeys;
 import org.compiere.model.I_C_Order;
 import org.compiere.model.I_M_PriceList;
 import org.compiere.util.TimeUtil;
